+++ example/cradle.py
@@ -1,14 +1,7 @@
 import cadquery as cq
 from cqspoolterrain import Spool,Cradle
 
-
 bp_spool = Spool()
-#bp.height = 100
-#bp.radius = 100
-#bp.wall_width = 3
-#bp.cut_radius = 40
-#bp.internal_wall_width = 4
-#bp.internal_z_translate = -3
 bp_spool.make()
 spool_ex = (
       bp_spool.build()
@@ -21,7 +14,6 @@ bp.height = bp_spool.radius - bp_spool.cut_radius+2
 bp.angle = 45
 bp.make(bp_spool)
 cradle_ex = bp.build().translate((0,0,bp.height/2))
-
 
 #show_object(cradle_ex)
 cq.exporters.export(cradle_ex,"stl/cradle.stl")